refactor: consolidate duplicate monitoring code
