+++ apps/portal/src/app/router.js
@@ -1,13 +1,24 @@
 import { Router as createRouter } from 'express';
 import { viewHomepage } from './views/home/controller.js';
-import { createRoutes as createMonitoringRoutes } from './monitoring.js';
+import { createMonitoringRoutes } from '@pins/crowndev-lib/controllers/monitoring.js';
+import { loadConfig } from './config.js';
+import { getLogger } from '#util/logger.js';
+import { getDatabaseClient } from '#util/database.js';
 
 /**
  * @returns {import('express').Router}
  */
 export function buildRouter() {
 	const router = createRouter();
-	const monitoringRoutes = createMonitoringRoutes();
+	const logger = getLogger();
+	const config = loadConfig();
+	const dbClient = getDatabaseClient();
+
+	const monitoringRoutes = createMonitoringRoutes({
+		config,
+		dbClient,
+		logger
+	});
 
 	router.use('/', monitoringRoutes);
